Split examples/fastapi_example.py into main.py + fastapi_routes.py
Separates app bootstrapping from route definitions so the demo-app repo's
sync process (diff lib-examples/fastapi_routes.py against its own routes
file) doesn't have to re-derive which lines are routes vs. app setup on
every sync.

- examples/fastapi_routes.py (renamed from fastapi_example.py, history
  preserved): APIRouter with every route, ContactForm/FeedbackForm,
  CSRF helpers, Jinja2Templates/safe_json_filter, FORM_REGISTRY, and
  create_refer_path. Dropped imports that moved out with app bootstrap
  (FastAPI, StaticFiles, SessionMiddleware, os, sys) and the now-unneeded
  sys.path.insert, since this module is only ever imported, never run
  directly.
- examples/main.py (new): thin composition root — FastAPI() construction,
  session middleware, the /static mount, app.include_router(router), and
  the uvicorn entrypoint banner.

Updated every reference to match: makefile (ex-run/ex-test/kill, using
`main` instead of `fastapi_example`), tests/test_integration.py and
tests/test_layouts.py (which monkeypatch templates/safe_json_filter for
test-time template resolution — repointed to fastapi_routes, the module
that now actually owns those names), README.md (+ docs/index.md, synced
the same way `make create-docs-local` does), docs/validation_guide.md,
and examples/README.md.

Verified both entrypoint styles work in a fresh subprocess: `python
examples/main.py` (direct execution) and `cd examples && python -c
"import main"` (the flat-import style `uvicorn main:app` relies on).
